Move all frontend inside v1 directory
