+++ src/v1/components/main/main.component.jsx
