+++ src/v1/components/skeleton/skeleton.component.jsx
