+++ src/v1/components/tour/singleTour.component.jsx
